2/send feature
