+++ package.json
@@ -139,7 +139,7 @@
     "cpx": "=1.5.0",
     "css-hot-loader": "1.4.3",
     "css-loader": "0.28.7",
-    "cypress": "3.3.1",
+    "cypress": "3.3.2",
     "cypress-cucumber-preprocessor": "1.12.0",
     "cypress-pipe": "1.3.3",
     "del-cli": "1.1.0",
